+++ contrib/opencensus-ext-flask/opencensus/ext/flask/flask_middleware.py
@@ -156,9 +156,6 @@ class FlaskMiddleware(object):
                 HTTP_PATH, flask.request.path
             )
             tracer.add_attribute_to_current_span(
-                HTTP_ROUTE, flask.request.path
-            )
-            tracer.add_attribute_to_current_span(
                 HTTP_URL, str(flask.request.url)
             )
             execution_context.set_opencensus_attr(
@@ -180,8 +177,11 @@ class FlaskMiddleware(object):
         try:
             tracer = execution_context.get_opencensus_tracer()
             tracer.add_attribute_to_current_span(
+                HTTP_ROUTE, flask.request.url_rule.rule
+            )
+            tracer.add_attribute_to_current_span(
                 HTTP_STATUS_CODE,
-                str(response.status_code)
+                response.status_code
             )
         except Exception:  # pragma: NO COVER
             log.error('Failed to trace request', exc_info=True)