+++ contrib/opencensus-ext-requests/tests/test_requests_trace.py
@@ -15,10 +15,12 @@
 import unittest
 
 import mock
+import requests
 from opencensus.trace.tracers import noop_tracer
 
 from opencensus.ext.requests import trace
 from opencensus.trace import span as span_module, execution_context
+from opencensus.trace import status as status_module
 
 
 class Test_requests_trace(unittest.TestCase):
@@ -95,19 +97,30 @@ class Test_requests_trace(unittest.TestCase):
 
         wrapped = trace.wrap_requests(mock_func)
 
-        url = 'http://localhost:8080'
+        url = 'http://localhost:8080/test'
 
         with patch, patch_thread:
             wrapped(url)
 
-        expected_attributes = {'http.url': url, 'http.status_code': '200'}
-        expected_name = '[requests]get'
+        expected_attributes = {
+            'http.host': 'localhost:8080',
+            'http.method': 'GET',
+            'http.path': '/test',
+            'http.status_code': 200,
+            'http.url': url,
+        }
+        expected_name = '/test'
+        expected_status = status_module.Status(0)
 
         self.assertEqual(span_module.SpanKind.CLIENT,
                          mock_tracer.current_span.span_kind)
         self.assertEqual(expected_attributes,
                          mock_tracer.current_span.attributes)
         self.assertEqual(expected_name, mock_tracer.current_span.name)
+        self.assertEqual(
+            expected_status.__dict__,
+            mock_tracer.current_span.status.__dict__
+        )
 
     def test_wrap_requests_blacklist_ok(self):
         mock_return = mock.Mock()
@@ -138,7 +151,7 @@ class Test_requests_trace(unittest.TestCase):
         with patch_tracer, patch_attr, patch_thread:
             wrapped(url)
 
-        expected_name = '[requests]get'
+        expected_name = '/'
 
         self.assertEqual(expected_name, mock_tracer.current_span.name)
 
@@ -204,6 +217,149 @@ class Test_requests_trace(unittest.TestCase):
 
         self.assertEqual(None, mock_tracer.current_span)
 
+    def test_wrap_requests_timeout(self):
+        mock_return = mock.Mock()
+        mock_return.status_code = 200
+        return_value = mock_return
+        mock_func = mock.Mock()
+        mock_func.__name__ = 'get'
+        mock_func.return_value = return_value
+        mock_func.side_effect = requests.Timeout
+        mock_tracer = MockTracer()
+
+        patch = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'get_opencensus_tracer',
+            return_value=mock_tracer)
+
+        patch_thread = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'is_exporter',
+            return_value=False)
+
+        wrapped = trace.wrap_requests(mock_func)
+
+        url = 'http://localhost:8080/test'
+
+        with patch, patch_thread:
+            with self.assertRaises(requests.Timeout):
+                wrapped(url)
+
+        expected_attributes = {
+            'http.host': 'localhost:8080',
+            'http.method': 'GET',
+            'http.path': '/test',
+            'http.url': url,
+        }
+        expected_name = '/test'
+        expected_status = status_module.Status(4, 'request timed out')
+
+        self.assertEqual(span_module.SpanKind.CLIENT,
+                         mock_tracer.current_span.span_kind)
+        self.assertEqual(expected_attributes,
+                         mock_tracer.current_span.attributes)
+        self.assertEqual(expected_name, mock_tracer.current_span.name)
+        self.assertEqual(
+            expected_status.__dict__,
+            mock_tracer.current_span.status.__dict__
+        )
+
+    def test_wrap_requests_invalid_url(self):
+        mock_return = mock.Mock()
+        mock_return.status_code = 200
+        return_value = mock_return
+        mock_func = mock.Mock()
+        mock_func.__name__ = 'get'
+        mock_func.return_value = return_value
+        mock_func.side_effect = requests.URLRequired
+        mock_tracer = MockTracer()
+
+        patch = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'get_opencensus_tracer',
+            return_value=mock_tracer)
+
+        patch_thread = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'is_exporter',
+            return_value=False)
+
+        wrapped = trace.wrap_requests(mock_func)
+
+        url = 'http://localhost:8080/test'
+
+        with patch, patch_thread:
+            with self.assertRaises(requests.URLRequired):
+                wrapped(url)
+
+        expected_attributes = {
+            'http.host': 'localhost:8080',
+            'http.method': 'GET',
+            'http.path': '/test',
+            'http.url': url,
+        }
+        expected_name = '/test'
+        expected_status = status_module.Status(3, 'invalid URL')
+
+        self.assertEqual(span_module.SpanKind.CLIENT,
+                         mock_tracer.current_span.span_kind)
+        self.assertEqual(expected_attributes,
+                         mock_tracer.current_span.attributes)
+        self.assertEqual(expected_name, mock_tracer.current_span.name)
+        self.assertEqual(
+            expected_status.__dict__,
+            mock_tracer.current_span.status.__dict__
+        )
+        self.assertRaises(requests.URLRequired, mock_func)
+
+    def test_wrap_requests_exception(self):
+        mock_return = mock.Mock()
+        mock_return.status_code = 200
+        return_value = mock_return
+        mock_func = mock.Mock()
+        mock_func.__name__ = 'get'
+        mock_func.return_value = return_value
+        mock_func.side_effect = requests.TooManyRedirects
+        mock_tracer = MockTracer()
+
+        patch = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'get_opencensus_tracer',
+            return_value=mock_tracer)
+
+        patch_thread = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'is_exporter',
+            return_value=False)
+
+        wrapped = trace.wrap_requests(mock_func)
+
+        url = 'http://localhost:8080/test'
+
+        with patch, patch_thread:
+            with self.assertRaises(requests.TooManyRedirects):
+                wrapped(url)
+
+        expected_attributes = {
+            'http.host': 'localhost:8080',
+            'http.method': 'GET',
+            'http.path': '/test',
+            'http.url': url,
+        }
+        expected_name = '/test'
+        expected_status = status_module.Status(2, '')
+
+        self.assertEqual(span_module.SpanKind.CLIENT,
+                         mock_tracer.current_span.span_kind)
+        self.assertEqual(expected_attributes,
+                         mock_tracer.current_span.attributes)
+        self.assertEqual(expected_name, mock_tracer.current_span.name)
+        self.assertEqual(
+            expected_status.__dict__,
+            mock_tracer.current_span.status.__dict__
+        )
+        self.assertRaises(requests.TooManyRedirects, mock_func)
+
     def test_wrap_session_request(self):
         wrapped = mock.Mock(return_value=mock.Mock(status_code=200))
 
@@ -220,16 +376,25 @@ class Test_requests_trace(unittest.TestCase):
             'is_exporter',
             return_value=False)
 
-        url = 'http://localhost:8080'
+        url = 'http://localhost:8080/test'
         request_method = 'POST'
         kwargs = {}
 
         with patch, patch_thread:
-            trace.wrap_session_request(wrapped, 'Session.request',
-                                       (request_method, url), kwargs)
+            trace.wrap_session_request(
+                wrapped, 'Session.request',
+                (request_method, url), kwargs
+            )
 
-        expected_attributes = {'http.url': url, 'http.status_code': '200'}
-        expected_name = '[requests]POST'
+        expected_attributes = {
+            'http.host': 'localhost:8080',
+            'http.method': 'POST',
+            'http.path': '/test',
+            'http.status_code': 200,
+            'http.url': url,
+        }
+        expected_name = '/test'
+        expected_status = status_module.Status(0)
 
         self.assertEqual(span_module.SpanKind.CLIENT,
                          mock_tracer.current_span.span_kind)
@@ -237,6 +402,10 @@ class Test_requests_trace(unittest.TestCase):
                          mock_tracer.current_span.attributes)
         self.assertEqual(kwargs['headers']['x-trace'], 'some-value')
         self.assertEqual(expected_name, mock_tracer.current_span.name)
+        self.assertEqual(
+            expected_status.__dict__,
+            mock_tracer.current_span.status.__dict__
+        )
 
     def test_wrap_session_request_blacklist_ok(self):
         def wrapped(*args, **kwargs):
@@ -261,14 +430,16 @@ class Test_requests_trace(unittest.TestCase):
             'is_exporter',
             return_value=False)
 
-        url = 'http://localhost'
+        url = 'http://localhost/'
         request_method = 'POST'
 
         with patch_tracer, patch_attr, patch_thread:
-            trace.wrap_session_request(wrapped, 'Session.request',
-                                       (request_method, url), {})
+            trace.wrap_session_request(
+                wrapped, 'Session.request',
+                (request_method, url), {}
+            )
 
-        expected_name = '[requests]POST'
+        expected_name = '/'
         self.assertEqual(expected_name, mock_tracer.current_span.name)
 
     def test_wrap_session_request_blacklist_nok(self):
@@ -296,8 +467,11 @@ class Test_requests_trace(unittest.TestCase):
         request_method = 'POST'
 
         with patch_tracer, patch_attr, patch_thread:
-            trace.wrap_session_request(wrapped, 'Session.request',
-                                       (request_method, url), {})
+            trace.wrap_session_request(
+                wrapped, 'Session.request',
+                (request_method, url), {}
+            )
+
         self.assertEqual(None, mock_tracer.current_span)
 
     def test_wrap_session_request_exporter_thread(self):
@@ -325,8 +499,11 @@ class Test_requests_trace(unittest.TestCase):
         request_method = 'POST'
 
         with patch_tracer, patch_attr, patch_thread:
-            trace.wrap_session_request(wrapped, 'Session.request',
-                                       (request_method, url), {})
+            trace.wrap_session_request(
+                wrapped, 'Session.request',
+                (request_method, url), {}
+            )
+
         self.assertEqual(None, mock_tracer.current_span)
 
     def test_header_is_passed_in(self):
@@ -349,8 +526,10 @@ class Test_requests_trace(unittest.TestCase):
         kwargs = {}
 
         with patch, patch_thread:
-            trace.wrap_session_request(wrapped, 'Session.request',
-                                       (request_method, url), kwargs)
+            trace.wrap_session_request(
+                wrapped, 'Session.request',
+                (request_method, url), kwargs
+            )
 
         self.assertEqual(kwargs['headers']['x-trace'], 'some-value')
 
@@ -374,8 +553,10 @@ class Test_requests_trace(unittest.TestCase):
         kwargs = {'headers': {'key': 'value'}}
 
         with patch, patch_thread:
-            trace.wrap_session_request(wrapped, 'Session.request',
-                                       (request_method, url), kwargs)
+            trace.wrap_session_request(
+                wrapped, 'Session.request',
+                (request_method, url), kwargs
+            )
 
         self.assertEqual(kwargs['headers']['key'], 'value')
         self.assertEqual(kwargs['headers']['x-trace'], 'some-value')
@@ -401,10 +582,156 @@ class Test_requests_trace(unittest.TestCase):
         kwargs = {'headers': {'x-trace': 'original-value'}}
 
         with patch, patch_thread:
-            trace.wrap_session_request(wrapped, 'Session.request',
-                                       (request_method, url), kwargs)
+            trace.wrap_session_request(
+                wrapped, 'Session.request',
+                (request_method, url), kwargs
+            )
 
         self.assertEqual(kwargs['headers']['x-trace'], 'some-value')
+
+    def test_wrap_session_request_timeout(self):
+        wrapped = mock.Mock(return_value=mock.Mock(status_code=200))
+        wrapped.side_effect = requests.Timeout
+
+        mock_tracer = MockTracer(
+            propagator=mock.Mock(
+                to_headers=lambda x: {'x-trace': 'some-value'}))
+
+        patch = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'get_opencensus_tracer',
+            return_value=mock_tracer)
+        patch_thread = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'is_exporter',
+            return_value=False)
+
+        url = 'http://localhost:8080/test'
+        request_method = 'POST'
+        kwargs = {}
+
+        with patch, patch_thread:
+            with self.assertRaises(requests.Timeout):
+                trace.wrap_session_request(
+                    wrapped, 'Session.request',
+                    (request_method, url), kwargs
+                )
+
+        expected_attributes = {
+            'http.host': 'localhost:8080',
+            'http.method': 'POST',
+            'http.path': '/test',
+            'http.url': url,
+        }
+        expected_name = '/test'
+        expected_status = status_module.Status(4, 'request timed out')
+
+        self.assertEqual(span_module.SpanKind.CLIENT,
+                         mock_tracer.current_span.span_kind)
+        self.assertEqual(expected_attributes,
+                         mock_tracer.current_span.attributes)
+        self.assertEqual(kwargs['headers']['x-trace'], 'some-value')
+        self.assertEqual(expected_name, mock_tracer.current_span.name)
+        self.assertEqual(
+            expected_status.__dict__,
+            mock_tracer.current_span.status.__dict__
+        )
+
+    def test_wrap_session_request_invalid_url(self):
+        wrapped = mock.Mock(return_value=mock.Mock(status_code=200))
+        wrapped.side_effect = requests.URLRequired
+
+        mock_tracer = MockTracer(
+            propagator=mock.Mock(
+                to_headers=lambda x: {'x-trace': 'some-value'}))
+
+        patch = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'get_opencensus_tracer',
+            return_value=mock_tracer)
+        patch_thread = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'is_exporter',
+            return_value=False)
+
+        url = 'http://localhost:8080/test'
+        request_method = 'POST'
+        kwargs = {}
+
+        with patch, patch_thread:
+            with self.assertRaises(requests.URLRequired):
+                trace.wrap_session_request(
+                    wrapped, 'Session.request',
+                    (request_method, url), kwargs
+                )
+
+        expected_attributes = {
+            'http.host': 'localhost:8080',
+            'http.method': 'POST',
+            'http.path': '/test',
+            'http.url': url,
+        }
+        expected_name = '/test'
+        expected_status = status_module.Status(3, 'invalid URL')
+
+        self.assertEqual(span_module.SpanKind.CLIENT,
+                         mock_tracer.current_span.span_kind)
+        self.assertEqual(expected_attributes,
+                         mock_tracer.current_span.attributes)
+        self.assertEqual(kwargs['headers']['x-trace'], 'some-value')
+        self.assertEqual(expected_name, mock_tracer.current_span.name)
+        self.assertEqual(
+            expected_status.__dict__,
+            mock_tracer.current_span.status.__dict__
+        )
+
+    def test_wrap_session_request_exception(self):
+        wrapped = mock.Mock(return_value=mock.Mock(status_code=200))
+        wrapped.side_effect = requests.TooManyRedirects
+
+        mock_tracer = MockTracer(
+            propagator=mock.Mock(
+                to_headers=lambda x: {'x-trace': 'some-value'}))
+
+        patch = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'get_opencensus_tracer',
+            return_value=mock_tracer)
+        patch_thread = mock.patch(
+            'opencensus.ext.requests.trace.execution_context.'
+            'is_exporter',
+            return_value=False)
+
+        url = 'http://localhost:8080/test'
+        request_method = 'POST'
+        kwargs = {}
+
+        with patch, patch_thread:
+            with self.assertRaises(requests.TooManyRedirects):
+                trace.wrap_session_request(
+                    wrapped, 'Session.request',
+                    (request_method, url), kwargs
+                )
+
+        expected_attributes = {
+            'http.host': 'localhost:8080',
+            'http.method': 'POST',
+            'http.path': '/test',
+            'http.url': url,
+        }
+        expected_name = '/test'
+        expected_status = status_module.Status(2, '')
+
+        self.assertEqual(span_module.SpanKind.CLIENT,
+                         mock_tracer.current_span.span_kind)
+        self.assertEqual(expected_attributes,
+                         mock_tracer.current_span.attributes)
+        self.assertEqual(kwargs['headers']['x-trace'], 'some-value')
+        self.assertEqual(expected_name, mock_tracer.current_span.name)
+        self.assertEqual(
+            expected_status.__dict__,
+            mock_tracer.current_span.status.__dict__
+        )
 
 
 class MockTracer(object):
@@ -414,8 +741,7 @@ class MockTracer(object):
         self.propagator = propagator
 
     def start_span(self):
-        span = mock.Mock()
-        span.attributes = {}
+        span = MockSpan()
         self.current_span = span
         return span
 
@@ -424,3 +750,11 @@ class MockTracer(object):
 
     def add_attribute_to_current_span(self, key, value):
         self.current_span.attributes[key] = value
+
+
+class MockSpan(object):
+    def __init__(self):
+        self.attributes = {}
+
+    def set_status(self, status):
+        self.status = status